(PC-23099)[PRO] refactor: clean offer status icons
