+++ pro/src/screens/OfferIndividual/Status/StatusToggleButton.tsx
@@ -3,7 +3,7 @@ import React, { useCallback } from 'react'
 import { api } from 'apiClient/api'
 import { OfferStatus } from 'apiClient/v1'
 import useNotification from 'hooks/useNotification'
-import { ReactComponent as StatusInactiveIcon } from 'icons/ico-status-inactive.svg'
+import { ReactComponent as FullHideIcon } from 'icons/full-hide.svg'
 import { ReactComponent as StrokeCheckIcon } from 'icons/stroke-check.svg'
 import { Button } from 'ui-kit'
 import { ButtonVariant } from 'ui-kit/Button/types'
@@ -42,9 +42,7 @@ const StatusToggleButton = ({
       variant={ButtonVariant.TERNARY}
       disabled={[OfferStatus.PENDING, OfferStatus.REJECTED].includes(status)}
       onClick={toggleOfferActiveStatus}
-      Icon={
-        status !== OfferStatus.INACTIVE ? StatusInactiveIcon : StrokeCheckIcon
-      }
+      Icon={status !== OfferStatus.INACTIVE ? FullHideIcon : StrokeCheckIcon}
     >
       {status !== OfferStatus.INACTIVE ? 'Désactiver' : 'Publier'}
     </Button>
